+++ docker-support/run.sh
@@ -2,9 +2,8 @@
 
 echo "Verifying env variables presence."
 declare -a REQUIRED_ENV_VARS=(
-                                "${PROFILE_MANAGER_CONNECTOR_BASE_URL}"
-                                "${TASK_MANAGER_CONNECTOR_BASE_URL}"
-                                "${HUB_CONNECTOR_BASE_URL}"
+                                "${PLATFORM_BASE_URL}"
+                                "${COMP_AUTH_KEY}"
                               )
 
 for e in "${REQUIRED_ENV_VARS[@]}"