+++ src/wenet_service_api/api/ws/resource/common.py
@@ -8,8 +8,8 @@ from typing import List, Optional
 from flask import request
 from flask_restful import Resource, abort
 
-from wenet.common.model.app.app_dto import AppStatus, App
-from wenet.common.model.scope import Scope
+from wenet.model.scope import Scope
+from wenet.model.app import AppStatus, App
 from wenet_service_api.common.exception.exceptions import ResourceNotFound
 from wenet_service_api.connector.collector import ServiceConnectorCollector
 
@@ -194,7 +194,7 @@ class AuthenticatedResource(Resource):
         app_id = consumer_id.replace("app_", "")
 
         try:
-            app = self._service_connector_collector.hub_connector.get_app(app_id)
+            app = self._service_connector_collector.hub_connector.get_app_details(app_id)
         except ResourceNotFound:
             logger.info(f"Invalid app [{app_id}]")
             abort(403, message="Invalid app")
@@ -208,7 +208,7 @@ class AuthenticatedResource(Resource):
             logger.info("Using debug authentication")
             return Oauth2Result(authenticated_user_id, scopes, app)
 
-        if app.status == AppStatus.ACTIVE:
+        if app.status == AppStatus.STATUS_ACTIVE:
             return Oauth2Result(authenticated_user_id, scopes, app)
         else: