+++ tests/wenet_service_api_test/api/ws/resources/test_task_interface.py
@@ -5,8 +5,8 @@ import json
 from mock import Mock
 
 from tests.wenet_service_api_test.api.common.common_test_case import CommonTestCase
-from wenet.common.model.norm.norm import Norm, NormOperator
-from wenet.common.model.task.task import Task, TaskGoal
+from wenet.model.norm import Norm, NormOperator
+from wenet.model.task.task import Task, TaskGoal
 from wenet_service_api.api.ws.resource.common import WenetSource
 
 
@@ -90,7 +90,7 @@ class TestTaskInterface(CommonTestCase):
         )
 
         mock_put = Mock(return_value=task)
-        self.service_collector_connector.task_manager_connector.updated_task = mock_put
+        self.service_collector_connector.task_manager_connector.update_task = mock_put
 
         response = self.client.put("/task/%s" % task_id, json=task.to_repr(), headers={"apikey": self.AUTHORIZED_APIKEY, "x-wenet-source": WenetSource.COMPONENT.value})
         self.assertEqual(response.status_code, 200)