+++ tests/wenet_service_api_test/api/ws/resources/test_task_transaction_interface.py
@@ -3,7 +3,7 @@ from __future__ import absolute_import, annotations
 from mock import Mock
 
 from tests.wenet_service_api_test.api.common.common_test_case import CommonTestCase
-from wenet.common.model.task.transaction import TaskTransaction
+from wenet.model.task.transaction import TaskTransaction
 from wenet_service_api.api.ws.resource.common import WenetSource
 
 
@@ -15,7 +15,7 @@ class TestTaskTransactionInterface(CommonTestCase):
 
         mock_post = Mock()
 
-        self.service_collector_connector.task_manager_connector.post_task_transaction = mock_post
+        self.service_collector_connector.task_manager_connector.create_task_transaction = mock_post
 
         response = self.client.post("/task/transaction", json=task_transaction.to_repr(), headers={"apikey": self.AUTHORIZED_APIKEY, "x-wenet-source": WenetSource.COMPONENT.value})
         self.assertEqual(response.status_code, 201)